refactored the files into separate folders moved the type definitions for customers and payment request into it type file in its folder
