+++ lib/customer/customer.test.ts
@@ -1,5 +1,5 @@
 import { test, mock, expect } from "bun:test"
-import { Paystack } from "./paystack"
+import { Paystack } from "../paystack"
 
 
 const paystack = mock(() => new Paystack(import.meta.env.PAYSTACK_SECRET_KEY ?? ""))
